Add idle-timeout/idle-timeout-command flags
This allows running an arbitrary shell command when a specified timeout
duration expires with no keypress events.  The primary intended use-case is
for engaging a sleep or hibernate mode if the screen has been locked and
the keyboard untouched for a given time.

The implementation generalizes the existing discard_passwd_timeout
mechanism to make the duration configurable via a command-line flag and
allow specifying a command to be run when the timer expires.